Add Dockerfile for working in a docker container
Note that I chose to add Datomic's repo to the project.clj config, which requires that we explicitly define a version dependency for it to fetch.

It also requires setting up credentials with Leiningen. Note that the approach I'm taking here is totally up for discussion, but I wanted to keep it simple while allowing for the Datomic installation to occur during Docker-build. The idea is that you'll follow the "Full Disk Encryption" option mentioned in Leiningen's DEPLOY doc, to set up an `:auth` profile:

```clojure
;; ./.lein_profiles.clj
{:auth {:repository-auth {#"my\.datomic\.com" {:username "..."
                                               :password "..."}}}}
```
This will then be installed as `/root/.lein/profiles.clj` in the Docker image on build, so your credentials are available to lein later during the `lein deps` command.

See here for more about this approach: https://github.com/technomancy/leiningen/blob/master/doc/DEPLOY.md#full-disk-encryption

The preferred approach is using a GPG-encrypted file, but this approach dramatically complicates use with Docker. This plain-text approach is simpler but does have some security implications, of course. However, given that we're all using full-disk encryption anyway (I hope) and the credential we're storing is (ideally) tied to a Datomic-Pro-Starter license rather than an account where a real Datomic purchase was made, it seems acceptable.

The `./.lein_profiles.clj` profile file could also be a hard-link (symlinks don't work unfortunately) to ~/.lein/profiles.clj so it's easy to share with local dev as well as other Docker builds using this pattern...